One-table verbs from dplyr for keyed_df

Defined methods for dplyr generic single table functions. Most of them preserve 'keyed_df' class and 'keys' attribute (excluding summarise with scoped variants, distinct and do which remove them). Also these methods modify rows in keys according to the rows modification in reference data frame (if any).

# S3 method for keyed_df
select(.tbl, ...)

# S3 method for keyed_df
rename(.tbl, ...)

# S3 method for keyed_df
mutate(.tbl, ...)

# S3 method for keyed_df
transmute(.tbl, ...)

# S3 method for keyed_df
summarise(.tbl, ...)

# S3 method for keyed_df
group_by(.tbl, ...)

# S3 method for keyed_df
ungroup(.tbl, ...)

# S3 method for keyed_df
rowwise(.tbl)

# S3 method for keyed_df
distinct(.tbl, ..., .keep_all = FALSE)

# S3 method for keyed_df
do(.tbl, ...)

# S3 method for keyed_df
arrange(.tbl, ..., .by_group = FALSE)

# S3 method for keyed_df
filter(.data, ...)

# S3 method for keyed_df
slice(.tbl, ...)

Arguments

.tbl, .data

A keyed object.

...

Appropriate arguments for functions.

.keep_all

Parameter for dplyr::distinct.

.by_group

Parameter for dplyr::arrange.

Details

dplyr::transmute() is supported implicitly with dplyr::mutate() support.

dplyr::rowwise() is not supposed to be generic in dplyr. Use rowwise.keyed_df directly.

All scoped variants of present functions are also supported.

See also

Examples

mtcars %>% key_by(vs, am) %>% dplyr::mutate(gear = 1)
#> # A keyed object. Keys: vs, am 
#>                      mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> Mazda RX4           21.0   6 160.0 110 3.90 2.620 16.46  0  1    1    4
#> Mazda RX4 Wag       21.0   6 160.0 110 3.90 2.875 17.02  0  1    1    4
#> Datsun 710          22.8   4 108.0  93 3.85 2.320 18.61  1  1    1    1
#> Hornet 4 Drive      21.4   6 258.0 110 3.08 3.215 19.44  1  0    1    1
#> Hornet Sportabout   18.7   8 360.0 175 3.15 3.440 17.02  0  0    1    2
#> Valiant             18.1   6 225.0 105 2.76 3.460 20.22  1  0    1    1
#> Duster 360          14.3   8 360.0 245 3.21 3.570 15.84  0  0    1    4
#> Merc 240D           24.4   4 146.7  62 3.69 3.190 20.00  1  0    1    2
#> Merc 230            22.8   4 140.8  95 3.92 3.150 22.90  1  0    1    2
#> Merc 280            19.2   6 167.6 123 3.92 3.440 18.30  1  0    1    4
#> Merc 280C           17.8   6 167.6 123 3.92 3.440 18.90  1  0    1    4
#> Merc 450SE          16.4   8 275.8 180 3.07 4.070 17.40  0  0    1    3
#> Merc 450SL          17.3   8 275.8 180 3.07 3.730 17.60  0  0    1    3
#> Merc 450SLC         15.2   8 275.8 180 3.07 3.780 18.00  0  0    1    3
#> Cadillac Fleetwood  10.4   8 472.0 205 2.93 5.250 17.98  0  0    1    4
#> Lincoln Continental 10.4   8 460.0 215 3.00 5.424 17.82  0  0    1    4
#> Chrysler Imperial   14.7   8 440.0 230 3.23 5.345 17.42  0  0    1    4
#> Fiat 128            32.4   4  78.7  66 4.08 2.200 19.47  1  1    1    1
#> Honda Civic         30.4   4  75.7  52 4.93 1.615 18.52  1  1    1    2
#> Toyota Corolla      33.9   4  71.1  65 4.22 1.835 19.90  1  1    1    1
#> Toyota Corona       21.5   4 120.1  97 3.70 2.465 20.01  1  0    1    1
#> Dodge Challenger    15.5   8 318.0 150 2.76 3.520 16.87  0  0    1    2
#> AMC Javelin         15.2   8 304.0 150 3.15 3.435 17.30  0  0    1    2
#> Camaro Z28          13.3   8 350.0 245 3.73 3.840 15.41  0  0    1    4
#> Pontiac Firebird    19.2   8 400.0 175 3.08 3.845 17.05  0  0    1    2
#> Fiat X1-9           27.3   4  79.0  66 4.08 1.935 18.90  1  1    1    1
#> Porsche 914-2       26.0   4 120.3  91 4.43 2.140 16.70  0  1    1    2
#> Lotus Europa        30.4   4  95.1 113 3.77 1.513 16.90  1  1    1    2
#> Ford Pantera L      15.8   8 351.0 264 4.22 3.170 14.50  0  1    1    4
#> Ferrari Dino        19.7   6 145.0 175 3.62 2.770 15.50  0  1    1    6
#> Maserati Bora       15.0   8 301.0 335 3.54 3.570 14.60  0  1    1    8
#> Volvo 142E          21.4   4 121.0 109 4.11 2.780 18.60  1  1    1    2